Added new enemy and fixed sticky keys bug

+++ Game Tutorial/src/com/tutorial/main/FastEnemy.java
@@ -19,7 +19,7 @@ public class FastEnemy extends GameObject{
 	}
 	
 	public Rectangle getBounds() {	// hit box
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 
 	
@@ -41,7 +41,7 @@ public class FastEnemy extends GameObject{
 	
 	public void render(Graphics g) {
 		g.setColor(Color.CYAN);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 		
 	}
